+++ main.py
@@ -1,99 +1,199 @@
-import requests, threading
+from __future__ import annotations
 
-from py_basic_commands  import create_file_dir, func_timer, join_path, get_dir_path_for_file, fprint
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from py_basic_commands  import try_listdir
 from selectolax.parser  import HTMLParser
+from send2trash         import send2trash
+from aiohttp            import ClientSession
+from fsutil             import join_filepath
+from tqdm               import tqdm
+from os                 import makedirs
+
+import httpx, asyncio
 
 
-def get_html_from_url(url):
-    resp = requests.get(url)
-    if resp.status_code != 200:
-        raise ValueError('Website error!')
 
-    html = HTMLParser(resp.text)
+def create_and_remove_empty_dir(func):
+    """Decorator to create and remove empty directories.
     
-    return html
+    Parameters
+    ----------
+    func : function
+        The function to be decorated.
+        
+    Returns
+    -------
+    function
+        The decorated function.
+    """
+    def wrapper(self, *args, **kwargs):
+        # Create the section directory
+        makedirs(self.section_dir, exist_ok=True)
 
-def threaded(fn):
-    def wrapper(*args, **kwargs):
-        thread = threading.Thread(target=fn, args=args, kwargs=kwargs)
-        thread.start()
-        return thread
+        # Run the function in an asyncio event loop
+        asyncio.run(func(self, *args, **kwargs))
+
+        # Remove the section directory if it's empty
+        if try_listdir(self.section_dir) == []:
+            send2trash(self.section_dir)
     return wrapper
 
 
+
 class Section:
-    def __init__(self, root, name:str, url_suffix:str) -> None:
+    def __init__(
+            self,
+            parent_self:StalenhagDownloader,
+            name:str,
+            section_href:str
+        ):
+        
+        self.parent_self = parent_self
         self.section_name = name.strip()
-        self.url_fill = root.website_homepage + '{}'
-        self.section_url = self.url_fill.format(url_suffix)
-        self.section_dir = join_path(root.main_dir_path, self.section_name)
-        self.section_file_path_fill = join_path(self.section_dir, '{}')
+        self.section_url = self.parent_self.url_lambda(section_href)
+        self.section_dir = join_filepath(self.parent_self.main_dir_path, self.section_name)
 
-    @threaded
-    def get_section_images(self):
-        html = get_html_from_url(self.section_url)
-        self.href_lst = [x.attributes['href'] for x in html.css('[target=_blank]') if 'big' in x.attributes['href']]
 
-    def download_section_images(self):
-        @threaded
-        def download_image(href):
-            img_file_path = self.section_file_path_fill.format(get_dir_path_for_file(href, 'fnam'))
-            img_url = self.url_fill.format(href)
+    async def fetch_section_image_urls(self, session:ClientSession):
+        """Fetch all image urls from the section.
+        
+        Parameters
+        ----------
+        session : ClientSession
+            The aiohttp ClientSession object.
+            
+        Returns
+        -------
+        list
+            The list of image urls.
+        """
+        img_url_lst = []
 
-            reqs = requests.get(img_url)
+        async with session.get(self.section_url) as resp:
+            for node in HTMLParser(await resp.text()).css('a[target="_blank"]'):
+                # Get the image href
+                img_href = node.attributes.get('href', '')
 
-            if reqs.status_code != 200:
-                return
+                # Only yield big images
+                if 'big' not in img_href: # type: ignore
+                    continue
+                
+                # Format the whole image url
+                img_url = self.parent_self.url_lambda(img_href)
+                if img_url not in img_url_lst:
+                    img_url_lst.append(img_url)
+                
+        return img_url_lst
 
-            with open(img_file_path, 'wb') as handler:
-                handler.write(reqs.content)
 
-        if not self.href_lst:
-            return
+    @create_and_remove_empty_dir
+    async def download_all_section_images(self):
+        """Download all images from the section to the section directory.
+        
+        Raises
+        ------
+        ValueError
+            If the response status is not 200.
+        """
+        async def download_image(img_url:str):
+            """Download an image from the section.
+            
+            Parameters
+            ----------
+            img_url : str
+                The image url.
+            """
+            async with session.get(img_url) as resp:
+                # Raise ValueError if response status is not 200
+                if resp.status != 200:
+                    raise ValueError(f'Error: {resp.status} -> {img_url}')
 
-        create_file_dir('d', self.section_dir)
+                # Save image to file in section directory
+                file_name = f'{self.section_name}_{img_url.split("/")[-1]}'
+                with open(
+                    join_filepath(self.section_dir, file_name),
+                    'wb'
+                    ) as handler:
+                    handler.write(await resp.read())
 
-        fprint(f'Downloading: {self.section_name}')
-        download_thr_lst = [download_image(href) for href in self.href_lst]
-        [thr.join() for thr in download_thr_lst]
+        async with ClientSession() as session:
+            # Download all big images from the section
+            await asyncio.gather(*[
+                download_image(img_url) for img_url in
+                await self.fetch_section_image_urls(session)
+            ])
+        
+
+
 
 
 class StalenhagDownloader:
-    """
-    The StalenhagDownloader class is used to download all images from the website simonstalenhag.se.
-    This class uses the 'requests' library to make HTTP GET requests to the website and 'selectolax' library to parse the HTML content of the website.
-    It uses threading to make request and download concurrently for faster downloading time.
-    It takes an optional argument 'main_dir_path' which represents the path where the downloaded images will be stored.
+    website_homepage = 'https://www.simonstalenhag.se'
 
-    Args:
-    - `main_dir_path` (str, optional): path where the downloaded images will be stored. Defaults to 'Stalenhag collection'.
-
-    Methods:
-    - `start()`: Initiates the download process.
-
-    """
-    website_homepage = 'https://www.simonstalenhag.se/'
-
-    def __init__(self, main_dir_path:str='Stalenhag collection'):
-        if not main_dir_path:
-            main_dir_path = 'Stalenhag collection'
-
-        self.main_dir_path = main_dir_path
-
-    @func_timer()
-    def start(self):
-        create_file_dir('d', self.main_dir_path)
-
-        html = get_html_from_url(self.website_homepage)
-        obj_lst = [Section(self, x.text(), x.attributes['href']) for x in html.css('span.style2 a')]
-
-        obj_thr_lst = [x.get_section_images() for x in obj_lst]
-        [x.join() for x in obj_thr_lst]
-
-        [x.download_section_images() for x in obj_lst]
-
-        print('Code complete!')
+    def __init__(self, dir_path:str='Stalenhag collection'):
+        self.main_dir_path = dir_path
+        self.url_lambda = lambda url_suffix : f'{self.website_homepage}/{url_suffix}'
 
 
-if __name__ == '__main__': 
+    @staticmethod
+    def get_html_from_url(url:str) -> HTMLParser:
+        """Get HTML from a website and return it as a HTMLParser object.
+        
+        Parameters
+        ----------
+        url : str
+            The website url.
+            
+        Returns
+        -------
+        HTMLParser
+            The HTMLParser object of the website.
+        """
+        resp = httpx.get(url)
+        if resp.status_code != 200:
+            raise ValueError('Website error!')
+
+        return HTMLParser(resp.text)
+
+
+    def start(self) -> None:
+        """Start the download of all images from the website."""
+        thread_lst = []
+
+
+        with ThreadPoolExecutor(max_workers=5) as executor:
+            for x in self.get_html_from_url(self.website_homepage).css('span[class="style2"] a'):
+                section_href = x.attributes['href']
+
+                if section_href is None:
+                    raise ValueError(f'No href found! -> {x.text()}')
+
+                thread_lst.append(
+                    executor.submit(
+                        Section(
+                            self,
+                            x.text(),
+                            section_href
+                        ).download_all_section_images
+                    )
+                )
+
+            print('Starting threads...')
+            pbar = tqdm(
+                total = len(thread_lst),
+                desc = 'Threads completed',
+                unit = 'thread'
+            )
+            for future in as_completed(thread_lst):
+                pbar.update(1)
+            pbar.close()
+
+        print('All threads completed!')
+
+        
+            
+
+
+
+if __name__ == '__main__':
     StalenhagDownloader().start()